+++ Zapt/View/FASearchBarWithActivity.m
@@ -76,7 +76,7 @@
     self.startCount -= 1;
 }
 
-- (int)startCount
+- (NSInteger)startCount
 {
     return _startCount;
 }
